pyproject.toml
[project]
name = "jira-tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jira>=3.10.5",
    "langchain>=0.3.27",
    "python-dotenv>=1.1.1",
    "mcp[cli]>=1.2.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/"]
